+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonDataGridViewDateTimePickerCell.cs
@@ -39,6 +39,7 @@ public class KryptonDataGridViewDateTimePickerCell : DataGridViewTextBoxCell
     private Size _calendarDimensions;
     private string _calendarTodayText;
     private Day _calendarFirstDayOfWeek;
+    private MonthCalendarView _calendarView;
     private bool _calendarShowToday;
     private bool _calendarCloseOnTodayClick;
     private bool _calendarShowTodayCircle;
@@ -65,6 +66,7 @@ public class KryptonDataGridViewDateTimePickerCell : DataGridViewTextBoxCell
         _calendarDimensions = new Size(1, 1);
         _calendarTodayText = "Today:";
         _calendarFirstDayOfWeek = Day.Default;
+        _calendarView = MonthCalendarView.Days;
         _calendarShowToday = true;
         _calendarCloseOnTodayClick = false;
         _calendarShowTodayCircle = true;
@@ -109,6 +111,7 @@ public class KryptonDataGridViewDateTimePickerCell : DataGridViewTextBoxCell
         dateTimeCell.CalendarDimensions = CalendarDimensions;
         dateTimeCell.CalendarTodayText = CalendarTodayText;
         dateTimeCell.CalendarFirstDayOfWeek = CalendarFirstDayOfWeek;
+        dateTimeCell.CalendarView = CalendarView;
         dateTimeCell.CalendarShowToday = CalendarShowToday;
         dateTimeCell.CalendarCloseOnTodayClick = CalendarCloseOnTodayClick;
         dateTimeCell.CalendarShowTodayCircle = CalendarShowTodayCircle;
@@ -349,6 +352,24 @@ public class KryptonDataGridViewDateTimePickerCell : DataGridViewTextBoxCell
     }
 
     /// <summary>
+    /// The CalendarView property replicates the one from the KryptonDateTimePicker control
+    /// </summary>
+    [DefaultValue(MonthCalendarView.Days)]
+    public MonthCalendarView CalendarView
+    {
+        get => _calendarView;
+
+        set
+        {
+            if (_calendarView != value)
+            {
+                SetCalendarView(RowIndex, value);
+                OnCommonChange();
+            }
+        }
+    }
+
+    /// <summary>
     /// The CalendarShowToday property replicates the one from the KryptonDateTimePicker control
     /// </summary>
     [DefaultValue(true)]
@@ -489,6 +510,7 @@ public class KryptonDataGridViewDateTimePickerCell : DataGridViewTextBoxCell
                 dateTime.CalendarDimensions = CalendarDimensions;
                 dateTime.CalendarTodayText = CalendarTodayText;
                 dateTime.CalendarFirstDayOfWeek = CalendarFirstDayOfWeek;
+                dateTime.CalendarView = CalendarView;
                 dateTime.CalendarShowToday = CalendarShowToday;
                 dateTime.CalendarCloseOnTodayClick = CalendarCloseOnTodayClick;
                 dateTime.CalendarShowTodayCircle = CalendarShowTodayCircle;
@@ -1011,6 +1033,15 @@ DataGridView!.EditingControl as KryptonDataGridViewDateTimePickerEditingControl 
         if (OwnsEditingDateTimePicker(rowIndex))
         {
             EditingDateTimePicker.CalendarFirstDayOfWeek = value;
+        }
+    }
+
+    internal void SetCalendarView(int rowIndex, MonthCalendarView value)
+    {
+        _calendarView = value;
+        if (OwnsEditingDateTimePicker(rowIndex))
+        {
+            EditingDateTimePicker.CalendarView = value;
         }
     }